requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
langchain>=0.1.0
langgraph>=0.0.20
langchain-google-genai>=0.0.6
langchain-community>=0.0.10
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-search-results>=2.4.0
python-dotenv>=1.0.0 